requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.3.0
blinker==1.7.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.8
Flask==3.0.1
Flask-Cors==4.0.0
h11==0.14.0
httptools==0.6.4
idna==3.10
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
Mako==1.3.10
MarkupSafe==2.1.4
mock-open==1.4.0
mysql-connector-python==8.3.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
passlib==1.7.4
pluggy==1.5.0
psycopg==3.2.9
pydantic==2.10.6
pydantic-settings==2.10.1
pydantic_core==2.27.2
PyJWT==2.10.1
pytest==8.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.45.3
threadpoolctl==3.4.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
Werkzeug==3.0.1
